refactor(util): add check node and remove unused methods
